requirements.txt
langchain-openai
browser_use
python-dotenv
playwright